debian: Set execute bit on rules file
It is expected that debian/rules is executable[1]. This is enforced at
various levels when building via sonic-buildimage so the sonic-bmp
submodule often ends up with a lone modified content entry because
something has "fixed" the file to have the required permissions. Commit
this mode change so that future builds don't end up with an unwanted
modified content on sonic-bmp.

[1] - https://www.man7.org/linux/man-pages/man5/deb-src-rules.5.html

Signed-off-by: Chris Packham <chris.packham@alliedtelesis.co.nz>